requirements.txt
# ── Telegram Bot framework ────────────────────────────────────────────────
python-telegram-bot==20.7

# ── HTTP та .env ──────────────────────────────────────────────────────────
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.9.0

# ── Планувальник ──────────────────────────────────────────────────────────
apscheduler>=3.10.4

# ── Google API (Sheets + Drive) ───────────────────────────────────────────
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
gspread>=5.12.0

# ── Браузер-автоматизація (scrapers) ──────────────────────────────────────
DrissionPage>=4.0.4

# ── AWS S3 + Textract (AI-sorter) ─────────────────────────────────────────
boto3>=1.34.0

# ── OpenAI (AI-sorter) ────────────────────────────────────────────────────
openai>=1.12.0

# ── Дані та зображення ────────────────────────────────────────────────────
pandas>=2.1.0
Pillow>=10.2.0

# ── OCR ───────────────────────────────────────────────────────────────────
paddleocr>=2.7.0
paddlepaddle>=2.6.0
pytesseract>=0.3.10
psd-tools>=1.9.31

# ── Утиліти ───────────────────────────────────────────────────────────────
python-dateutil>=2.8.2
